refactor & add utils function
